requirements.txt
certifi
charset-normalizer
idna
numpy
opencv-python
python-dotenv
requests
urllib3
torch
Pillow
matplotlib
PyYAML
scipy
tqdm
seaborn
opencv-python-headless
streamlit